--- src/main/java/com/kb04/starroad/Service/BoardService1.java
@@ -1,8 +1,0 @@
-package com.kb04.starroad.Service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class BoardService1 {
-
-}
