--- src/main/java/com/kb04/starroad/Service/BoardService3.java
@@ -1,13 +1,0 @@
-package com.kb04.starroad.Service;
-
-import com.kb04.starroad.Entity.Board;
-import com.kb04.starroad.Repository.BoardRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class BoardService3 {
-
-    @Autowired
-    private BoardRepository boardRepository;
-}
