+++ src/main/webapp/WEB-INF/views/member/member.jsp
@@ -8,7 +8,7 @@
     <script src="//code.jquery.com/jquery-3.6.0.min.js"></script>
     <script src="//code.jquery.com/jquery-latest.min.js"></script>
     <link rel="stylesheet" type="text/css" href="${path}/resources/static/css/common.css">
-    <link rel="stylesheet" type="text/css" href="${path}/resources/static/css/member.css">
+    <link rel="stylesheet" type="text/css" href="${path}/resources/static/css/member/member.css">
     <script type="text/javascript">
         $(function() {
             $("#navbar").load("${path}/resources/common_jsp/navbar.jsp");
