+++ src/main/webapp/WEB-INF/views/product/product_result.jsp
@@ -11,7 +11,7 @@
           integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
     <link rel="stylesheet" type="text/css" href="${path}/resources/static/css/common.css"/>
     <link rel="stylesheet" type="text/css" href="${path}/resources/static/css/nav.css">
-    <link rel="stylesheet" type="text/css" href="${path}/resources/static/css/product.css">
+    <link rel="stylesheet" type="text/css" href="${path}/resources/static/css/product/product.css">
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.0/jquery.min.js"></script>
     <script type="text/javascript">
         $(function () {
