+++ src/main/webapp/resources/common_jsp/password.jsp
@@ -10,7 +10,7 @@
     <link rel="stylesheet" href="${path}/resources/static/css/mypage/check_password.css">
     <script src="//code.jquery.com/jquery-3.6.0.min.js"></script>
     <script src="//code.jquery.com/jquery-latest.min.js"></script>
-    <link rel="stylesheet" type="text/css" href="${path}/resources/static/css/member.css">
+    <link rel="stylesheet" type="text/css" href="${path}/resources/static/css/member/member.css">
 
     <script type="text/javascript">
         $(function() {
